+++ source/renderer/zettlr-body.js
@@ -76,6 +76,17 @@ class ZettlrBody {
     global.notify = (msg) => {
       this.notify(msg)
     }
+
+    // Afterwards, activate the event listeners of the window controls
+    $('.windows-window-controls .minimise, .linux-window-controls .minimise').click((e) => {
+      global.ipc.send('win-minimise')
+    })
+    $('.windows-window-controls .resize, .linux-window-controls .maximise').click((e) => {
+      global.ipc.send('win-maximise')
+    })
+    $('.windows-window-controls .close, .linux-window-controls .close').click((e) => {
+      global.ipc.send('win-close')
+    })
   }
 
   /**