+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - Zettlr keeps some margin between the cursor where you are writing and the window edges, i.e. it won't touch the window edges anymore, but keep a nice distance.
 - Quicklook windows can now be "popped out" so that they are no longer bound to the main window but can be dragged onto different displays, etc.
 - Zettlr will now highlight YAML blocks at the beginning of Markdown files if they resemble the format of Pandoc metadata blocks.
+- Windows and Linux windows now follow macOS in having no native window frame, but instead they employ the same strategy as macOS: The toolbar is the top element inside the main window of Zettlr, featuring window controls and, additionally, a button to open the application menu from the toolbar.
 
 ## Under the hood